old-tests/webapps/WebStorage/tests/approved/storage_local_length.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Web Storage</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>storage_local_length</h3>
	</p>
	
	<hr>

	<div id="log"></div>

	<script>
		if (('localStorage' in window) && window.localStorage !== null){
			localStorage.clear();

			test(function() {
				assert_equals(localStorage.length, 0, "localStorage.length")
			}, "localStorage length test after clear()")

			localStorage.setItem("name", "user1");
			localStorage.setItem("age", "20");

			test(function() {
				assert_equals(localStorage.length, 2, "localStorage.length")
			}, "localStorage length test after setItem()")

		} else {
			test(function()
			{
				assert_unreached("localStorage is not supported.");
			}, "localStorage support test")
		}	
	</script>

 </body>
</html>
